wip: checkpoint before i do something stupid

+++ models/driving_modules/model_lstm_driving_global.py
@@ -9,7 +9,7 @@ from utils.args import DefaultArguments
 
 cuda = True if torch.cuda.is_available() else False
 device = torch.device("cuda:0" if cuda else "cpu")
-FloatTensor = torch.cuda.FloatTensor if cuda else torch.FloatTensor
+FloatTensor = torch.cuda.BFloat16Tensor if cuda else torch.FloatTensor
 LongTensor = torch.cuda.LongTensor if cuda else torch.LongTensor
